auth/requirements.txt
fastapi
pydantic[email]
SQLAlchemy
uvicorn
requests
psycopg2
python-jose
passlib[bcrypt]
